viaiMcpRag/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber==0.11.6
python-docx>=1.0.0
openpyxl>=3.1.0
fastapi>=0.109.0
motor>=3.3.0
pymongo>=4.6.0
python-multipart>=0.0.6
faiss-cpu>=1.7.4
uvicorn>=0.27.0
gunicorn>=21.2.0
python-dotenv
numpy==1.26.4
